pyproject.toml
[project]
name = "notify-me-sempai"
version = "0.1.0"
description = ""
authors = [
    { name = "DmitriiDolinskii", email = "dmitriy.d.dolinskiy@gmail.com" },
]
readme = "README.md"
requires-python = "=3.12.7"
dependencies = ["aio-pika (>=9.5.4,<10.0.0)", "websockets (>=14.2,<15.0)", "pyjwt (>=2.10.1,<3.0.0)"]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
